First commit
This is a screenplay to video converter.
You create your characters, then you add them to the scene, you choose
a voice for each character, then you type up a screenplay. then it all
gets converted into a video you can upload to  youtube!

This is a Visual Studio project,  which does run on mono on Linux, but
for some reason it has bugs on Mac.